requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
requests
python-multipart
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
unstructured
nbformat
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext[openai]==0.7.5